refactor: re-structure ui layer
