+++ pull_request_manager/lib/ui/screens/pull_request/widgets/detail_item.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class PullRequestDetailItem extends StatelessWidget {
+class DetailItem extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
